+++ src/UnityFx.Async/Implementation/Promises/RebindResult{T,U}.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Alexander Bogarsukov.
+﻿// Copyright (c) 2018-2020 Alexander Bogarsukov.
 // Licensed under the MIT license. See the LICENSE.md file in the project root for more information.
 
 using System;
@@ -56,7 +56,7 @@ namespace UnityFx.Async.Promises
 							break;
 
 						case Func<T, U> f2:
-							TrySetResult(f2((op as IAsyncOperation<T>).Result));
+							TrySetResult(f2(((IAsyncOperation<T>)op).Result));
 							break;
 
 						default: